Use `u32` to store transaction version
Use a type consistent with how the version is specified.

+++ zebra-consensus/src/error.rs
@@ -48,7 +48,7 @@ pub enum TransactionError {
     WrongVersion,
 
     #[error("transaction version {0} not supported by the network upgrade {1:?}")]
-    UnsupportedByNetworkUpgrade(usize, zebra_chain::parameters::NetworkUpgrade),
+    UnsupportedByNetworkUpgrade(u32, zebra_chain::parameters::NetworkUpgrade),
 
     #[error("must have at least one input: transparent, shielded spend, or joinsplit")]
     NoInputs,
